Near completion

+++ src/Item.java
@@ -2,10 +2,10 @@ public class Item {
 
     private String itemName;
     private double price;
-    private int calories;
+    private double calories;
 
     // constructor for the Item class
-    public Item(String itemName, double price, int calories) {
+    public Item(String itemName, double price, double calories) {
         this.itemName = itemName;
         this.price = price;
         this.calories = calories;
@@ -24,7 +24,7 @@ public class Item {
     }
 
 
-    public int getCalories() {
+    public double getCalories() {
         return calories;
     }
